refactor: update file structure, moved Component
