+++ src/components/ui/FormItemInput.tsx
@@ -1,4 +1,4 @@
-import { Input } from "@/shared/components/ui/input";
+import { Input } from "@/app/components/ui/input";
 import React from "react";
 
 interface FormItemInputProps {
