+++ src/components/ui/command.tsx
@@ -11,7 +11,7 @@ import {
   DialogDescription,
   DialogHeader,
   DialogTitle,
-} from "@/shared/components/ui/dialog"
+} from "@/app/components/ui/dialog"
 
 function Command({
   className,
